requirements.txt
jupyter
matplotlib
numpy
Pillow
torch
torchvision
gle-code @ git+https://github.com/unibe-cns/gle-code
